Fixed timer countdown and adding implementation for multiple laps

+++ proj/src/view/game.h
@@ -41,7 +41,7 @@ typedef struct Game {
     int player_turn_input_sign; // -1 for left, 0 for none, 1 for right
     bool pause_requested;
 
-    int timer_count_down;
+    float timer_count_down;
 
 } Game;
